requirements.txt
numpy>=1.26
torch>=2.1
pyyaml>=6.0
